+++ Source/DREAM3DLib/IOFilters/util/SurfaceDataContainerWriter.cpp
@@ -350,7 +350,7 @@ int SurfaceDataContainerWriter::writeMeshData(hid_t dcGid)
     int32_t rank = 2; // THIS NEEDS TO BE THE SAME AS THE NUMBER OF ELEMENTS IN THE Structure from SurfaceMesh::DataStruc
     hsize_t dims[2] = {facesPtr->getNumberOfTuples(), 3};
 
-    size_t* data = reinterpret_cast<size_t*>(facesPtr->getPointer(0));
+    int32_t* data = reinterpret_cast<int32_t*>(facesPtr->getPointer(0));
 
     err = QH5Lite::writePointerDataset(dcGid, DREAM3D::HDF5::FacesName, rank, dims, data);
     if (err < 0)
